# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FleetSaaS school bus tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stations:
    get:
      operationId: listStations
      tags: [stations]
      summary: List all geofenced stations
      responses:
        "200":
          description: Station list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Station"
    post:
      operationId: createStation
      tags: [stations]
      summary: Add a new geofenced station (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Station"
  /stations/{id}:
    delete:
      operationId: deleteStation
      tags: [stations]
      summary: Delete a station
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DRIVERS ────────────────────────────────────────────────────────────────
...
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List all routes for this tenant
      responses:
        "200":
          description: Route list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
    post:
      operationId: createRoute
      tags: [routes]
      summary: Create a new route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "201":
          description: Created route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
  /routes/{id}:
    patch:
      operationId: updateRoute
      tags: [routes]
      summary: Update route name or assigned driver/vehicle (fleet swap)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteUpdate"
      responses:
        "200":
          description: Updated route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Delete a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /routes/{id}/stations:
    get:
      operationId: listRouteStations
      tags: [routes]
      summary: Get ordered stations for a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ordered station list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteStation"
    post:
      operationId: addRouteStation
      tags: [routes]
      summary: Add a station to a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteStationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteStation"
  /routes/{id}/stations/{stationId}:
    delete:
      operationId: removeRouteStation
      tags: [routes]
      summary: Remove a station from a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: stationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    Route:
      type: object
      required: [id, tenantId, name, isActive]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        driverId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        driverName:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
    RouteInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        driverId:
          type: integer
        vehicleId:
          type: integer
    RouteUpdate:
      type: object
      properties:
        name:
          type: string
        driverId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        isActive:
          type: boolean
    RouteStation:
      type: object
      required: [id, routeId, stationId, position]
      properties:
        id:
          type: integer
        routeId:
          type: integer
        stationId:
          type: integer
        position:
          type: integer
        stationName:
          type: string
        lat:
          type: number
        lng:
          type: number
        radius:
          type: integer
    RouteStationInput:
      type: object
      required: [stationId]
      properties:
        stationId:
          type: integer
        position:
          type: integer
...
    Station:
      type: object
      required: [id, name, lat, lng]
      properties:
        id:
          type: integer
        name:
          type: string
        lat:
          type: number
        lng:
          type: number
        radius:
          type: integer
        tenantId:
          type: integer
    StationInput:
      type: object
      required: [name, lat, lng]
      properties:
        name:
          type: string
        lat:
          type: number
        lng:
          type: number
        radius:
          type: integer
...